requirements.txt
asgiref==3.8.1
Django==5.1.3
djangorestframework==3.15.2
mysqlclient==2.2.6
PyMySQL==1.1.1
sqlparse==0.5.2
tzdata==2024.2
